app/requirements.txt
fastapi==0.95.1
uvicorn==0.23.1
streamlit==1.17.0
requests==2.28.2
spacy==3.5.2
pdfplumber==0.7.8
python-docx==0.8.11
python-multipart==0.0.5
PyPDF2==1.26.0
pandas==1.5.3
scikit-learn==1.2.2
